add mapping from streaming models 2 db models [WIP]

+++ modules/db-writer/src/main/scala/fi/spectrumlabs/db/writer/models/cardano/PoolEvent.scala
@@ -5,7 +5,7 @@ import derevo.derive
 
 @derive(encoder, decoder)
 final case class PoolEvent(
-  outCollateral: Int,
+  outCollateral: Long,
   poolCoinLq: CoinWrapper,
   poolCoinX: CoinWrapper,
   poolCoinY: CoinWrapper,
